+++ src/components/TodoCreate.jsx
@@ -33,7 +33,7 @@ const CircleButton = styled.button`
   justify-content: center;
 
   transition: 0.125s all ease-in;
-  ${props =>
+  ${(props) =>
     props.open &&
     css`
       background: #ff6b6b;
@@ -77,43 +77,14 @@ const Input = styled.input`
 `;
 
 function TodoCreate() {
-  const [open, setOpen] = useState(false);
-  const [value, setValue] = useState('');
-  const dispatch = useTodoDispatch();
-  const nextId = useTodoNextId();
-
-  const onToggle = () => setOpen(!open);
-  const onChange = e => setValue(e.target.value);
-  const onSubmit = e => {
-    e.preventDefault();
-    dispatch({
-      type: 'CREATE',
-      todo: {
-        id: nextId.current,
-        text: value,
-        done: false
-      }
-    });
-    nextId.current += 1;
-    setOpen(false);
-    setValue('');
-  };
-
   return (
     <>
-      {open && (
-        <InsertFormPositioner>
-          <InsertForm onSubmit={onSubmit}>
-            <Input
-              autoFocus
-              onChange={onChange}
-              value={value}
-              placeholder="할 일을 입력 후, Enter 를 누르세요"
-            />
-          </InsertForm>
-        </InsertFormPositioner>
-      )}
-      <CircleButton onClick={onToggle} open={open}>
+      <InsertFormPositioner>
+        <InsertForm>
+          <Input autoFocus placeholder="할 일을 입력 후, Enter 를 누르세요" />
+        </InsertForm>
+      </InsertFormPositioner>
+      <CircleButton>
         <MdAdd />
       </CircleButton>
     </>
