+++ src/components/TodoItem.jsx
@@ -39,7 +39,7 @@ const CheckCircle = styled.div`
   justify-content: center;
   margin-right: 20px;
   cursor: pointer;
-  ${props =>
+  ${(props) =>
     props.done &&
     css`
       border: 1px solid #38d9a9;
@@ -51,37 +51,21 @@ const Text = styled.div`
   flex: 1;
   font-size: 21px;
   color: #495057;
-  ${props =>
+  ${(props) =>
     props.done &&
     css`
       color: #ced4da;
     `}
 `;
 
-function TodoItem({ id, done, text }) {
-  const dispatch = useTodoDispatch();
-
-  const onToggle = () => {
-    dispatch({
-      type: 'TOGGLE',
-      id
-    });
-  };
-
-  const onRemove = () => {
-    dispatch({
-      type: 'REMOVE',
-      id
-    });
-  };
-
+function TodoItem({ text }) {
   return (
     <TodoItemBlock>
-      <CheckCircle done={done} onClick={onToggle}>
-        {done && <MdDone />}
+      <CheckCircle>
+        <MdDone />
       </CheckCircle>
-      <Text done={done}>{text}</Text>
-      <Remove onClick={onRemove}>
+      <Text>{text}</Text>
+      <Remove>
         <MdDelete />
       </Remove>
     </TodoItemBlock>
